requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
chromadb==0.5.5
sentence-transformers==3.1.1
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
